+++ include/borsh.h
@@ -19,100 +19,11 @@
 #ifndef BORSH_CPP20_H
 #define BORSH_CPP20_H
 
-#include <array>
-#include <cstddef>
-#include <cstring>
-#include <memory>
-#include <vector>
-#include <type_traits>
-#include <string>
-#include <stdexcept>
-#include <bit>
-#include <cstdint>
-#include <algorithm>
-#include <vector>
-#include <cmath>
-#include <memory>
-
-namespace borsh
-{
-
-auto serialize(ArrayType auto (&array)[], Serializer& serializer)
-{
-    return serializer(array);
-}
-
-auto serialize(ScalarType auto& value, Serializer& serializer)
-{
-    return serializer(value);
-}
-
-auto serialize(SerializableVector auto& value, Serializer& serializer)
-{
-    return serializer(value);
-}
-
-template <typename T>
-    requires ScalarType<T> || ScalarArrayType<T>
-std::vector<uint8_t> serialize(const T& value)
-{
-    std::vector<uint8_t> buffer;
-    to_bytes(value, buffer);
-    return buffer;
-}
-
-template <SerializableNonScalar T> std::vector<uint8_t> serialize(T& object)
-{
-    std::vector<uint8_t> buffer;
-    const uint8_t*       data = buffer.data();
-    Serializer           serializer(buffer, data, SerializerDirection::Serialize);
-    serialize(object, serializer);
-    return buffer;
-}
-
-std::vector<uint8_t> serialize(SerializableNonScalarArray auto (&array)[])
-{
-    std::vector<uint8_t> buffer;
-    const uint8_t*       data = buffer.data();
-    Serializer           serializer(buffer, data, SerializerDirection::Serialize);
-    serialize(array, serializer);
-    return buffer;
-}
-
-template <typename T>
-    requires ScalarType<T>
-T deserialize(std::vector<uint8_t>& buffer)
-{
-    const uint8_t* data = buffer.data();
-    T              value;
-    from_bytes(value, data);
-    return value;
-}
-
-template <typename T, std::size_t N>
-    requires ScalarType<T>
-void deserialize(T (&value)[N], std::vector<uint8_t>& buffer)
-{
-    const uint8_t* data = buffer.data();
-    from_bytes(value, data);
-}
-
-template <SerializableNonScalar T> T deserialize(std::vector<uint8_t>& buffer)
-{
-    const uint8_t* data = buffer.data();
-    auto           object = T{};
-    Serializer     serializer(buffer, data, SerializerDirection::Deserialize);
-    serialize(object, serializer);
-    return object;
-}
-
-void deserialize(SerializableNonScalarArray auto (&value)[], std::vector<uint8_t>& buffer)
-{
-    const uint8_t* data = buffer.data();
-    Serializer     serializer(buffer, data, SerializerDirection::Deserialize);
-    serialize(value, serializer);
-}
-
-} // namespace borsh
+#include "borsh/concepts.h"
+#include "borsh/utils.h"
+#include "borsh/converters.h"
+#include "borsh/serializer.h"
+#include "borsh/templates.h"
+#include "boost/ut.hpp"
 
 #endif
